+++ src/app/delivery/rest/search/search-api.ts
@@ -1,5 +1,5 @@
 import { Observable } from "rxjs";
 
 export interface SearchApi {
-  getSearchResult<T extends string>(arg: T, arg2: T): Observable<any>;
+  // getSearchResult<T extends string>(arg: T, arg2: T): Observable<any>;
 }
