+++ src/app/delivery/rest/search/search-result-mapper.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@angular/core';
-import { SearchResult } from "../../../shared/domain/search-result.model";
-import {MovieSearchResultListDto} from "./movie-search-result-list.dto";
-import {MoviePreviewDto} from "../movie/movie-preview.dto";
+import { MovieSearchResultListDto } from "./movie-search-result-list.dto";
+import { MoviePreviewDto } from "../movie/movie-preview.dto";
 
 @Injectable()
 export class SearchResultMapper {
-  public fromApi(searchResultDto: MovieSearchResultListDto): SearchResult[] {
+  // TODO
+/*  public fromApi(searchResultDto: MovieSearchResultListDto): SearchResult[] {
     return searchResultDto.results.map((dto: MoviePreviewDto) => {
       return {
         id: dto.externalId,
@@ -15,5 +15,5 @@ export class SearchResultMapper {
         releaseDate: dto.releaseDate
       };
     });
-  }
+  }*/
 }
