+++ src/app/delivery/rest/search/search.service.ts
@@ -2,7 +2,6 @@ import { Inject, Injectable } from '@angular/core';
 import { SearchApi } from "./search-api";
 import { HttpClient } from "@angular/common/http";
 import { map, Observable } from "rxjs";
-import { SearchResult } from "../../../shared/domain/search-result.model";
 import { SearchResultMapper } from "./search-result-mapper.service";
 import { GATEWAY_BASE_URL } from "../../../config/gateway.config";
 import { MovieSearchResultListDto } from "./movie-search-result-list.dto";
@@ -20,11 +19,12 @@ export class SearchService implements SearchApi {
               @Inject(GATEWAY_BASE_URL) private readonly endpoint: string) {
   }
 
-  public getSearchResult(searchTerm: string, page?: string): Observable<SearchResult[]> {
-    return this.http.get<MovieSearchResultListDto>(
-        `${this.endpoint}/api/v3/search` + `?queryTerm=${searchTerm}` + `&page=${this.setGivenOrDefaultPage(page)}`,
-    ).pipe(map(response => this.mapper.fromApi(response)));
-  }
+  // TODO
+  /*  public getSearchResult(searchTerm: string, page?: string): Observable<SearchResult[]> {
+      return this.http.get<MovieSearchResultListDto>(
+          `${this.endpoint}/api/v3/search` + `?queryTerm=${searchTerm}` + `&page=${this.setGivenOrDefaultPage(page)}`,
+      ).pipe(map(response => this.mapper.fromApi(response)));
+    }*/
 
   private setGivenOrDefaultPage(page: string | undefined): string {
     if (page !== undefined) {
