+++ src/app/presentation/search-bar/search-bar.component.ts
