+++ src/app/presentation/search-result/search-result.component.html
@@ -1,3 +1,4 @@
+<!-- TODO
 <div class="card card-side bg-base-100 shadow-xl max-w-[90vw] min-w-[90vw] mb-4">
   <figure class="max-w-[200px] min-w-[200px]" >
     <img [src]="movie.imageUrl" alt="Movie"/>
@@ -15,3 +16,4 @@
     </div>
   </div>
 </div>
+-->
