+++ src/app/presentation/search-result/search-result.component.spec.ts
