--- src/app/shared/domain/search-result.model.ts
@@ -1,7 +1,0 @@
-export interface SearchResult {
-    "id": number,
-    "title": string,
-    "imageUrl": string,
-    "genres": string[],
-    "releaseDate": string,
-}
