+++ src/app/shared/shared.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SearchBarComponent } from './presentation/search-bar/search-bar.component';
+import { SearchBarComponent } from '../presentation/search-bar/search-bar.component';
 import { FormsModule } from "@angular/forms";
 import { SearchService } from "../delivery/rest/search/search.service";
 import { SearchResultMapper } from "../delivery/rest/search/search-result-mapper.service";
-import { SearchResultComponent } from './presentation/search-result/search-result.component';
+import { SearchResultComponent } from '../presentation/search-result/search-result.component';
 import { TruncatePipe } from './pipes/truncate.pipe';
 
 @NgModule({
